requirements.txt
streamlit==1.11.0
keras==2.10.0
tensorflow-gpu==2.10.0
numpy==1.24.4
Pillow==10.2.0